requirements.txt
# Django framework
Django>=4.2,<5.0

# PostgreSQL database
psycopg2-binary>=2.9
daphne==4.1.2

python-decouple
# Asynchronous task processing
dramatiq>=1.17

gunicorn>=20.1.0


# Redis broker for Dramatiq
redis>=5.0.1
django-environ
channels-redis==4.2.0
djangorestframework==3.16.0

honcho>=2.0.0

# Django Channels for WebSockets
channels>=4.0

# JWT authentication
djangorestframework-simplejwt>=5.5.1
PyJWT==2.9.0

django-cors-headers==4.7.0

# Geocoding / distance calculations
geopy>=2.3

# Email support
django-anymail>=9.0  # Optional if using services like Mailgun or SendGrid
# Or use just the default smtplib (no extra package required)

# Environment variables
python-dotenv>=1.0

# Optional utilities
requests>=2.31

python-dotenv